ViewModelにコンストラクタ経由で値を渡す方式に変更

+++ app/src/main/java/jp/ceed/android/mylapslogger/adatpter/PracticeResultsAdapter.kt
@@ -15,7 +15,7 @@ import jp.ceed.android.mylapslogger.dto.LapDto
 
 class PracticeResultsAdapter(
 	context: Context,
-	private val onClickListener: (Int) -> Unit
+	private val onClickListener: (Long) -> Unit
 ) : RecyclerView.Adapter<PracticeResultsAdapter.ViewHolder>() {
 
 	private val inflater: LayoutInflater = LayoutInflater.from(context)
@@ -45,7 +45,7 @@ class PracticeResultsAdapter(
 	}
 
 	private fun onClickSection(lapDto: LapDto){
-		if(lapDto.sessionId != 0){
+		if(lapDto.sessionId != 0L){
 			onClickListener(lapDto.sessionId)
 		}
 	}
